04_video/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
attrs==26.1.0
audioop-lts==0.2.2
certifi==2026.4.22
colorama==0.4.6
decorator==5.2.1
edge-tts==7.2.8
frozenlist==1.8.0
idna==3.13
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
moviepy==2.2.1
multidict==6.7.1
numpy==2.4.4
pillow==11.3.0
proglog==0.1.12
propcache==0.4.1
pydub==0.25.1
python-dotenv==1.2.2
srt==3.5.3
tabulate==0.10.0
tqdm==4.67.3
typing_extensions==4.15.0
yarl==1.23.0
